refactor(cart): extract useCart hook and update storefront pages
