+++ src/frontend/src/components/CartSummary.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { useCart } from '../context/CartContext';
+import { useCart } from '../context/useCart';
 
 export const CartSummary: React.FC = () => {
   const { items, subtotal, updateQuantity, removeItem } = useCart();
