+++ src/frontend/src/context/useCart.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { CartContext } from './CartContext';
+
+export const useCart = () => {
+  const context = useContext(CartContext);
+  if (!context) {
+    throw new Error('useCart must be used within a CartProvider');
+  }
+  return context;
+};
