+++ src/frontend/src/pages/Cart.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { useCart } from '../context/CartContext';
+import { useCart } from '../context/useCart';
 
 export const Cart: React.FC = () => {
   const { items, subtotal, updateQuantity, removeItem, clearCart } = useCart();
